refactor(app): promote VisualEffectView out of JumpMenuView
Sources/TBDApp/Panes/Transcript/Composer/CompletionOverlayView.swift
needs the same AppKit visual-effect bridge JumpMenuView already had
private to itself. Move it to Sources/TBDApp/Helpers/VisualEffectView.swift
as internal; the JumpMenu call site is unchanged (it already passed
blendingMode explicitly).
