+++ Sources/TBDApp/JumpMenu/JumpMenuView.swift
@@ -131,19 +131,3 @@ struct JumpMenuView: View {
             .padding(.vertical, 24)
     }
 }
-
-/// AppKit visual-effect bridge so the panel picks up the system menu material.
-private struct VisualEffectView: NSViewRepresentable {
-    let material: NSVisualEffectView.Material
-    let blendingMode: NSVisualEffectView.BlendingMode
-
-    func makeNSView(context: Context) -> NSVisualEffectView {
-        let v = NSVisualEffectView()
-        v.material = material
-        v.blendingMode = blendingMode
-        v.state = .active
-        return v
-    }
-
-    func updateNSView(_ nsView: NSVisualEffectView, context: Context) {}
-}
